+++ packages/cli/src/modules/services.js
@@ -139,19 +139,28 @@ export const ServicesModule = ({ config, profilePath }) => ({
         .option('from', { describe: 'Extension name', required: true })
         .option('service', { describe: 'Service to upgrade', required: true })
         .option('auth', { type: 'boolean', default: false, description: 'Authentication required' })
-        .option('dev', { type: 'boolean', default: false, description: 'Dev build' }),
+        .option('dev', { type: 'boolean', default: false, description: 'Dev build' })
+        .option('hot', { type: 'boolean', default: false, description: 'Hot upgrade' })
+        .option('name', { type: 'string', description: 'Container name' }),
 
       handler: asyncHandler(async argv => {
-        const { from: moduleName, service: serviceName, auth: authRequired, dev } = argv;
+        const { from: moduleName, service: serviceName, auth: authRequired, dev, hot } = argv;
 
         const service = getServiceInfo(moduleName, serviceName);
         const auth = authRequired ? getAuth(config, service) : undefined;
 
+        const { name = service.container_name } = argv;
+
         const { image: imageName } = service;
 
-        const container = await DockerContainer.find({ imageName });
-        if (container && container.started) {
-          throw new Error(`Unable to upgrade '${service.container_name}' while it's running.`);
+        const container = await DockerContainer.find({ imageName, dev });
+        const running = container?.started;
+
+        if (running) {
+          if (!hot) {
+            throw new Error(`Unable to upgrade '${service.container_name}' while it's running.`);
+          }
+          await container.stop();
         }
 
         // TODO(egorgripasov): Already up-to-date message.
@@ -159,6 +168,40 @@ export const ServicesModule = ({ config, profilePath }) => ({
         await dockerImage.pull(true);
 
         await DockerImage.cleanNotLatest(imageName);
+
+        if (running) {
+          const newContainer = await dockerImage.getOrCreateContainer({ name, restore: true });
+          await newContainer.start();
+        }
+      })
+    })
+
+    .command({
+      command: ['restore'],
+      describe: 'Restore service from saved state.',
+      builder: yargs => yargs
+        .option('from', { describe: 'Extension name', required: true })
+        .option('service', { describe: 'Service to upgrade', required: true })
+        .option('dev', { type: 'boolean', default: false, description: 'Dev build' })
+        .option('name', { type: 'string', description: 'Container name' }),
+
+      handler: asyncHandler(async argv => {
+        const { from: moduleName, service: serviceName, dev } = argv;
+
+        const service = getServiceInfo(moduleName, serviceName);
+
+        const { name = service.container_name } = argv;
+
+        const { image: imageName } = service;
+
+        const container = await DockerContainer.find({ imageName, dev });
+        if (container?.started) {
+          throw new Error(`Unable to restore '${service.container_name}' while it's running.`);
+        }
+
+        const dockerImage = new DockerImage({ service, dev });
+        const newContainer = await dockerImage.getOrCreateContainer({ name, restore: true });
+        await newContainer.start();
       })
     })
 
@@ -213,7 +256,7 @@ export const ServicesModule = ({ config, profilePath }) => ({
           ...additionalBinds
         ];
 
-        const container = await dockerImage.getOrCreateContainer(name, command, env, binds, hostNet, volumes);
+        const container = await dockerImage.getOrCreateContainer({ name, args: command, env, binds, hostNet, volumes });
         await container.start();
       })
     })